bug #298 Methods have been implemented in RenderingControl, RenderingControlImpl, RenderingControlProxy, also in DataManagementService, DMSAdapter. This new feature is added now but we will useful in the next version b/c we need to add new ST (server side). The SaveSetting button has been removed.
git-svn-id: file:///home/svn/shoola-omero/branches@1518 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ OME_2_2/SRC/org/openmicroscopy/shoola/agents/rnd/pane/QuantumPane.java
@@ -90,9 +90,9 @@ public class QuantumPane
 		int mini = (int) eventManager.getGlobalChannelWindowStart(INDEX);
 		int maxi = (int) eventManager.getGlobalChannelWindowEnd(INDEX);
 		int s = 
-			((Integer) eventManager.getChannelWindowStart(INDEX)).intValue();
+			(int) eventManager.getChannelWindowStart(INDEX);
 		int e = 
-			((Integer) eventManager.getChannelWindowEnd(INDEX)).intValue();
+			(int) eventManager.getChannelWindowEnd(INDEX);
 		codomainPane = new CodomainPane(eventManager.getRegistry(), manager);
 		domainPane = new DomainPane(eventManager.getRegistry(), manager, 
 									eventManager.getChannelData(), qDef, INDEX);
